+++ src/dmp/__tests__/commonStandard.spec.ts
@@ -1,5 +1,6 @@
 import fs from 'fs';
 import { Validator } from 'jsonschema';
+import { RDACommonStandardDMPType } from "../index";
 
 const SCHEMA_PATH = './schemas/dmp.schema.json';
 
@@ -8,7 +9,7 @@ describe('validate the RDA common standard', () => {
     const validator = new Validator();
     const schema = await fs.readFileSync(SCHEMA_PATH, 'utf8');
 
-    const dmp = {
+    const dmp: RDACommonStandardDMPType = {
       dmp: {
         title: 'Test DMP',
         dmp_id: {
@@ -22,10 +23,10 @@ describe('validate the RDA common standard', () => {
         contact: {
           name: 'Test Contact',
           mbox: 'tester@example.com',
-          contact_id: {
+          contact_id: [{
             identifier: '123456789',
             type: 'other'
-          }
+          }]
         },
         dataset: [{
           title: 'Test Dataset',
@@ -49,7 +50,7 @@ describe('validate the RDA common standard', () => {
     const validator = new Validator();
     const schema = await fs.readFileSync(SCHEMA_PATH, 'utf8');
 
-    const dmp = {
+    const dmp: RDACommonStandardDMPType = {
       dmp: {
         title: 'Test DMP',
         description: 'This is a test DMP',
@@ -66,10 +67,10 @@ describe('validate the RDA common standard', () => {
         contact: {
           name: 'Test Contact',
           mbox: 'tester@example.com',
-          contact_id: {
+          contact_id: [{
             identifier: 'https://orcid.org/0000-0000-0000-0000',
             type: 'orcid'
-          },
+          }],
           affiliation: [{
             name: 'Test University',
             affiliation_id: {
@@ -80,10 +81,10 @@ describe('validate the RDA common standard', () => {
         },
         contributor: [{
           name: 'Test Contact',
-          contributor_id: {
+          contributor_id: [{
             identifier: 'https://orcid.org/0000-0000-0000-0000',
             type: 'orcid'
-          },
+          }],
           affiliation: [{
             name: 'Test University',
             affiliation_id: {
@@ -117,10 +118,10 @@ describe('validate the RDA common standard', () => {
           metadata: [{
             description: 'Description of metadata',
             language: 'eng',
-            metadata_standard_id: {
+            metadata_standard_id: [{
               identifier: 'https://example.com/metadata-standards/123',
               type: 'url'
-            }
+            }]
           }],
           preservation_statement: 'Statement about preservation',
           security_and_privacy: [{
@@ -185,19 +186,23 @@ describe('validate the RDA common standard', () => {
       project: [{
         title: 'Test Project',
         description: 'This is a test project',
-        project_id: {
+        project_id: [{
           identifier: '123456789',
           type: 'other'
-        },
+        }],
         start: '2025-01-01',
         end: '2028-01-31',
         funding: [{
           name: 'Funder Organization',
-          funding_status: 'planned',
+          funding_status: 'granted',
           funder_id: {
             identifier: 'https://ror.org/0987654321',
             type: 'ror'
-          }
+          },
+          grant_id: [{
+            identifier: '123456789',
+            type: 'other'
+          }]
         }]
       }]
     };